refactor: Customer, User 패키지에 리팩토링
