+++ src/main/java/uos/uos25/common/DataInitializer.java
@@ -7,39 +7,39 @@ import jakarta.annotation.PostConstruct;
 import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
-import uos.uos25.customer.entity.Customer;
-import uos.uos25.customer.repository.CustomerRepository;
-import uos.uos25.disposal.entity.Disposal;
-import uos.uos25.disposal.repository.DisposalRepository;
-import uos.uos25.employee.entity.Employee;
-import uos.uos25.employee.entity.EmployeePosition;
-import uos.uos25.employee.entity.EmployeeWorkingHistory;
-import uos.uos25.employee.entity.PartTime;
-import uos.uos25.employee.repository.EmployeeRepository;
-import uos.uos25.employee.repository.EmployeeWorkingHistoryRepository;
-import uos.uos25.event.entity.Event;
-import uos.uos25.event.repository.EventRepository;
+import uos.uos25.headQuarter.entity.Event;
 import uos.uos25.headQuarter.entity.HeadQuarter;
+import uos.uos25.headQuarter.entity.Product;
+import uos.uos25.headQuarter.repository.EventRepository;
 import uos.uos25.headQuarter.repository.HeadQuarterRepository;
-import uos.uos25.inventory.entity.Inventory;
-import uos.uos25.inventory.repository.InventoryRepository;
-import uos.uos25.orders.entity.Orders;
-import uos.uos25.orders.entity.OrdersStatus;
-import uos.uos25.orders.repository.OrdersRepository;
-import uos.uos25.product.entity.Product;
-import uos.uos25.product.repository.ProductRepository;
-import uos.uos25.receipt.entity.Receipt;
-import uos.uos25.receipt.entity.ReceiptDetail;
-import uos.uos25.receipt.repository.ReceiptDetailRepository;
-import uos.uos25.receipt.repository.ReceiptRepository;
-import uos.uos25.returns.entity.Returns;
-import uos.uos25.returns.entity.ReturnsStatus;
-import uos.uos25.returns.repository.ReturnsRepository;
+import uos.uos25.headQuarter.repository.ProductRepository;
+import uos.uos25.order.entity.Disposal;
+import uos.uos25.order.entity.Orders;
+import uos.uos25.order.entity.OrdersStatus;
+import uos.uos25.order.entity.Returns;
+import uos.uos25.order.entity.ReturnsStatus;
+import uos.uos25.order.repository.DisposalRepository;
+import uos.uos25.order.repository.OrdersRepository;
+import uos.uos25.order.repository.ReturnsRepository;
 import uos.uos25.shop.entity.Disbursement;
 import uos.uos25.shop.entity.DisbursementType;
+import uos.uos25.shop.entity.Inventory;
 import uos.uos25.shop.entity.Shop;
 import uos.uos25.shop.repository.DisbursementRepository;
+import uos.uos25.shop.repository.InventoryRepository;
 import uos.uos25.shop.repository.ShopRepository;
+import uos.uos25.user.entity.Customer;
+import uos.uos25.user.entity.Employee;
+import uos.uos25.user.entity.EmployeePosition;
+import uos.uos25.user.entity.EmployeeWorkingHistory;
+import uos.uos25.user.entity.PartTime;
+import uos.uos25.user.receipt.entity.Receipt;
+import uos.uos25.user.receipt.entity.ReceiptDetail;
+import uos.uos25.user.receipt.repository.ReceiptDetailRepository;
+import uos.uos25.user.receipt.repository.ReceiptRepository;
+import uos.uos25.user.repository.CustomerRepository;
+import uos.uos25.user.repository.EmployeeRepository;
+import uos.uos25.user.repository.EmployeeWorkingHistoryRepository;
 
 @Component
 @RequiredArgsConstructor
