+++ src/main/java/uos/uos25/user/dto/request/CustomerCreateRequestDTO.java
@@ -1,4 +1,4 @@
-package uos.uos25.customer.dto.request;
+package uos.uos25.user.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
