+++ src/main/java/uos/uos25/user/dto/response/CustomerGetResponseDTO.java
@@ -1,19 +1,19 @@
-package uos.uos25.customer.dto.response;
+package uos.uos25.user.dto.response;
 
 import java.time.LocalDateTime;
 
 import lombok.Data;
-import uos.uos25.customer.entity.Customer;
+import uos.uos25.user.entity.Customer;
 
 @Data
-public class CustomerUpdateResponseDTO {
+public class CustomerGetResponseDTO {
     private final String phoneNumber;
     private final String nickname;
     private final LocalDateTime joinDate;
     private final Integer mileage;
 
-    public static CustomerUpdateResponseDTO fromEntity(Customer customer) {
-        return new CustomerUpdateResponseDTO(
+    public static CustomerGetResponseDTO fromEntity(Customer customer) {
+        return new CustomerGetResponseDTO(
                 customer.getPhoneNumber(),
                 customer.getNickname(),
                 customer.getJoinDate(),
