+++ src/main/java/uos/uos25/user/exception/CustomerAlreadyExists.java
@@ -1,4 +1,4 @@
-package uos.uos25.customer.exception;
+package uos.uos25.user.exception;
 
 import uos.uos25.common.error.BusinessException;
 import uos.uos25.common.error.ErrorCode;
