+++ src/main/java/uos/uos25/user/repository/CustomerRepository.java
@@ -1,7 +1,7 @@
-package uos.uos25.customer.repository;
+package uos.uos25.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import uos.uos25.customer.entity.Customer;
+import uos.uos25.user.entity.Customer;
 
 public interface CustomerRepository extends JpaRepository<Customer, String> {}
